Update new precompiles to use uint8

+++ stylus-tools/precompiles/ArbOwner.sol
@@ -212,7 +212,7 @@ interface ArbOwner {
     /// @notice Sets the max amount of stylus contract fragments that can be used to deploy a stylus contract
     /// @notice Available in ArbOS version 60 and above
     function setMaxStylusContractFragments(
-        uint16 maxFragments
+        uint8 maxFragments
     ) external;
 
     /// Emitted when a successful call is made to this precompile
